+++ frontend/jupiter_front/package.json
@@ -9,6 +9,7 @@
     "axios": "^1.4.0",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
+    "react-kakao-maps-sdk": "^1.1.9",
     "react-scripts": "5.0.1",
     "web-vitals": "^2.1.4"
   },
